adds useSendMessage hook
